moved friend search code
